Small linix fixes

+++ tests/county.hpp
@@ -45,9 +45,9 @@ namespace std {
         [[nodiscard]] size_t operator()(const county &val) const noexcept {
             // Yeah, this is how boost does it.
             size_t seed{};
-            seed ^= val.a + 0x9e3779b9 + (seed << 6) + (seed >> 2);
-            seed ^= val.b + 0x9e3779b9 + (seed << 6) + (seed >> 2);
-            seed ^= val.c + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+            seed ^= static_cast<size_t>(val.a) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+            seed ^= static_cast<size_t>(val.b) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+            seed ^= static_cast<size_t>(val.c) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
             return seed;
         }
     };
